refactor: annotation controller 점진적 리팩토링
