+++ src/main/java/next/controller/CreateUserController.java
@@ -1,24 +1,27 @@
 package next.controller;
 
+import core.annotation.web.Controller;
+import core.annotation.web.RequestMapping;
+import core.annotation.web.RequestMethod;
 import core.db.DataBase;
-import core.mvc.asis.Controller;
+import core.mvc.ModelAndView;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import next.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-public class CreateUserController implements Controller {
+@Controller
+public class CreateUserController {
     private static final Logger log = LoggerFactory.getLogger(CreateUserController.class);
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+    @RequestMapping(value = "/users/create", method = RequestMethod.POST)
+    public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         User user = new User(req.getParameter("userId"), req.getParameter("password"), req.getParameter("name"),
                 req.getParameter("email"));
         log.debug("User : {}", user);
 
         DataBase.addUser(user);
-        return "redirect:/";
+        return new ModelAndView("redirect:/");
     }
 }
